Button fixed for pathway/1
